services/ocr/requirements.txt
# Dépendances — Service OCR
fastapi==0.115.11
uvicorn==0.34.0
pytesseract==0.3.13
Pillow==11.1.0
PyMuPDF==1.25.3
pdf2image==1.17.0
python-multipart==0.0.20
minio==7.2.15
structlog==25.1.0
httpx==0.28.1
pydantic==2.11.0
